# rest-server/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Paperless Rest Server",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/token" : {
      "post" : {
        "tags" : [ "Login" ],
        "operationId" : "GetToken",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserInfo"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserInfo"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserInfo"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
    "/api/groups" : {
      "get" : {
        "tags" : [ "Login" ],
        "operationId" : "GetGroups",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "page_size",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_12"
                },
                "examples" : {
                  "default" : {
                    "value" : "{\"count\":0,\"next\":null,\"previous\":null,\"all\":[],\"results\":[]}"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Login" ],
        "operationId" : "CreateGroup",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/api_groups_body"
              },
              "examples" : {
                "default" : {
                  "value" : "{\"name\":\"testgroup\",\"permissions\":[\"add_document\",\"view_document\",\"change_document\",\"delete_document\",\"add_tag\",\"view_tag\",\"change_tag\",\"delete_tag\",\"add_correspondent\",\"view_correspondent\",\"change_correspondent\",\"delete_correspondent\",\"add_documenttype\",\"view_documenttype\",\"change_documenttype\",\"delete_documenttype\",\"add_storagepath\",\"view_storagepath\",\"change_storagepath\",\"delete_storagepath\",\"add_savedview\",\"view_savedview\",\"change_savedview\",\"delete_savedview\",\"add_paperlesstask\",\"view_paperlesstask\",\"change_paperlesstask\",\"delete_paperlesstask\",\"add_uisettings\",\"view_uisettings\",\"change_uisettings\",\"delete_uisettings\",\"add_note\",\"view_note\",\"change_note\",\"delete_note\",\"add_mailaccount\",\"view_mailaccount\",\"change_mailaccount\",\"delete_mailaccount\",\"add_mailrule\",\"view_mailrule\",\"change_mailrule\",\"delete_mailrule\",\"add_user\",\"view_user\",\"change_user\",\"delete_user\",\"add_group\",\"view_group\",\"change_group\",\"delete_group\",\"add_logentry\",\"view_logentry\",\"change_logentry\",\"delete_logentry\"]}"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "x-content-type" : "application/json"
                },
                "examples" : {
                  "default" : {
                    "value" : "{\"id\":1,\"name\":\"testgroup\",\"permissions\":[\"add_logentry\",\"change_logentry\",\"delete_logentry\",\"view_logentry\",\"add_group\",\"change_group\",\"delete_group\",\"view_group\",\"add_user\",\"change_user\",\"delete_user\",\"view_user\",\"add_correspondent\",\"change_correspondent\",\"delete_correspondent\",\"view_correspondent\",\"add_document\",\"change_document\",\"delete_document\",\"view_document\",\"add_documenttype\",\"change_documenttype\",\"delete_documenttype\",\"view_documenttype\",\"add_note\",\"change_note\",\"delete_note\",\"view_note\",\"add_paperlesstask\",\"change_paperlesstask\",\"delete_paperlesstask\",\"view_paperlesstask\",\"add_savedview\",\"change_savedview\",\"delete_savedview\",\"view_savedview\",\"add_storagepath\",\"change_storagepath\",\"delete_storagepath\",\"view_storagepath\",\"add_tag\",\"change_tag\",\"delete_tag\",\"view_tag\",\"add_uisettings\",\"change_uisettings\",\"delete_uisettings\",\"view_uisettings\",\"add_mailaccount\",\"change_mailaccount\",\"delete_mailaccount\",\"view_mailaccount\",\"add_mailrule\",\"change_mailrule\",\"delete_mailrule\",\"view_mailrule\"]}"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{id}" : {
      "put" : {
        "tags" : [ "Login" ],
        "operationId" : "UpdateGroup",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/groups_id_body"
              },
              "examples" : {
                "default" : {
                  "value" : "{\"id\":1,\"name\":\"testgroup2\",\"permissions\":[\"add_logentry\",\"change_logentry\",\"delete_logentry\",\"view_logentry\",\"add_group\",\"change_group\",\"delete_group\",\"view_group\",\"add_user\",\"change_user\",\"delete_user\",\"view_user\",\"add_correspondent\",\"change_correspondent\",\"delete_correspondent\",\"view_correspondent\",\"add_document\",\"change_document\",\"delete_document\",\"view_document\",\"add_documenttype\",\"change_documenttype\",\"delete_documenttype\",\"view_documenttype\",\"add_note\",\"change_note\",\"delete_note\",\"view_note\",\"add_paperlesstask\",\"change_paperlesstask\",\"delete_paperlesstask\",\"view_paperlesstask\",\"add_savedview\",\"change_savedview\",\"delete_savedview\",\"view_savedview\",\"add_storagepath\",\"change_storagepath\",\"delete_storagepath\",\"view_storagepath\",\"add_tag\",\"change_tag\",\"delete_tag\",\"view_tag\",\"add_uisettings\",\"change_uisettings\",\"delete_uisettings\",\"view_uisettings\",\"add_mailaccount\",\"change_mailaccount\",\"delete_mailaccount\",\"view_mailaccount\",\"add_mailrule\",\"change_mailrule\",\"delete_mailrule\",\"view_mailrule\"],\"set_permissions\":[\"add_logentry\",\"change_logentry\",\"delete_logentry\",\"view_logentry\",\"add_group\",\"change_group\",\"delete_group\",\"view_group\",\"add_user\",\"change_user\",\"delete_user\",\"view_user\",\"add_correspondent\",\"change_correspondent\",\"delete_correspondent\",\"view_correspondent\",\"add_document\",\"change_document\",\"delete_document\",\"view_document\",\"add_documenttype\",\"change_documenttype\",\"delete_documenttype\",\"view_documenttype\",\"add_note\",\"change_note\",\"delete_note\",\"view_note\",\"add_paperlesstask\",\"change_paperlesstask\",\"delete_paperlesstask\",\"view_paperlesstask\",\"add_savedview\",\"change_savedview\",\"delete_savedview\",\"view_savedview\",\"add_storagepath\",\"change_storagepath\",\"delete_storagepath\",\"view_storagepath\",\"add_tag\",\"change_tag\",\"delete_tag\",\"view_tag\",\"add_uisettings\",\"change_uisettings\",\"delete_uisettings\",\"view_uisettings\",\"add_mailaccount\",\"change_mailaccount\",\"delete_mailaccount\",\"view_mailaccount\",\"add_mailrule\",\"change_mailrule\",\"delete_mailrule\",\"view_mailrule\"],\"permissions_form\":{\"set_permissions\":[\"add_logentry\",\"change_logentry\",\"delete_logentry\",\"view_logentry\",\"add_group\",\"change_group\",\"delete_group\",\"view_group\",\"add_user\",\"change_user\",\"delete_user\",\"view_user\",\"add_correspondent\",\"change_correspondent\",\"delete_correspondent\",\"view_correspondent\",\"add_document\",\"change_document\",\"delete_document\",\"view_document\",\"add_documenttype\",\"change_documenttype\",\"delete_documenttype\",\"view_documenttype\",\"add_note\",\"change_note\",\"delete_note\",\"view_note\",\"add_paperlesstask\",\"change_paperlesstask\",\"delete_paperlesstask\",\"view_paperlesstask\",\"add_savedview\",\"change_savedview\",\"delete_savedview\",\"view_savedview\",\"add_storagepath\",\"change_storagepath\",\"delete_storagepath\",\"view_storagepath\",\"add_tag\",\"change_tag\",\"delete_tag\",\"view_tag\",\"add_uisettings\",\"change_uisettings\",\"delete_uisettings\",\"view_uisettings\",\"add_mailaccount\",\"change_mailaccount\",\"delete_mailaccount\",\"view_mailaccount\",\"add_mailrule\",\"change_mailrule\",\"delete_mailrule\",\"view_mailrule\"]}}"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_13"
                },
                "examples" : {
                  "default" : {
                    "value" : "{\"id\":1,\"name\":\"testgroup2\",\"permissions\":[\"add_logentry\",\"change_logentry\",\"delete_logentry\",\"view_logentry\",\"add_group\",\"change_group\",\"delete_group\",\"view_group\",\"add_user\",\"change_user\",\"delete_user\",\"view_user\",\"add_correspondent\",\"change_correspondent\",\"delete_correspondent\",\"view_correspondent\",\"add_document\",\"change_document\",\"delete_document\",\"view_document\",\"add_documenttype\",\"change_documenttype\",\"delete_documenttype\",\"view_documenttype\",\"add_note\",\"change_note\",\"delete_note\",\"view_note\",\"add_paperlesstask\",\"change_paperlesstask\",\"delete_paperlesstask\",\"view_paperlesstask\",\"add_savedview\",\"change_savedview\",\"delete_savedview\",\"view_savedview\",\"add_storagepath\",\"change_storagepath\",\"delete_storagepath\",\"view_storagepath\",\"add_tag\",\"change_tag\",\"delete_tag\",\"view_tag\",\"add_uisettings\",\"change_uisettings\",\"delete_uisettings\",\"view_uisettings\",\"add_mailaccount\",\"change_mailaccount\",\"delete_mailaccount\",\"view_mailaccount\",\"add_mailrule\",\"change_mailrule\",\"delete_mailrule\",\"view_mailrule\"]}"
                  }
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Login" ],
        "operationId" : "DeleteGroup",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Success"
          }
        }
      }
    },
...
    "/api/logs/{id}" : {
      "get" : {
        "tags" : [ "Config" ],
        "operationId" : "GetLog",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  },
                  "x-content-type" : "application/json"
                },
                "examples" : {
                  "default" : {
                    "value" : "[\"[2023-09-08 12:03:51,945] [INFO] [paperless.management.consumer] Using inotify to watch directory for changes: /usr/src/paperless/consume\",\"[2023-09-08 12:04:53,939] [INFO] [paperless.consumer] Consuming test.pdf\",\"[2023-09-08 12:04:53,947] [DEBUG] [paperless.consumer] Detected mime type: application/pdf\",\"[2023-09-08 12:04:54,030] [DEBUG] [paperless.consumer] Parser: RasterisedDocumentParser\",\"[2023-09-08 12:04:54,036] [DEBUG] [paperless.consumer] Parsing test.pdf...\",\"[2023-09-08 12:04:55,095] [DEBUG] [paperless.parsing.tesseract] Calling OCRmyPDF with args: {'input_file': PosixPath('/tmp/paperless/paperless-ngxm31l7wip/test.pdf'), 'output_file': PosixPath('/tmp/paperless/paperless-iabw4ryn/archive.pdf'), 'use_threads': True, 'jobs': '1', 'language': 'eng', 'output_type': 'pdfa', 'progress_bar': False, 'skip_text': True, 'clean': True, 'deskew': True, 'rotate_pages': True, 'rotate_pages_threshold': 12.0, 'sidecar': PosixPath('/tmp/paperless/paperless-iabw4ryn/sidecar.txt')}\",\"[2023-09-08 12:04:59,229] [DEBUG] [paperless.parsing.tesseract] Incomplete sidecar file: discarding.\",\"[2023-09-08 12:04:59,460] [DEBUG] [paperless.consumer] Generating thumbnail for test.pdf...\",\"[2023-09-08 12:04:59,561] [DEBUG] [paperless.parsing] Execute: convert -density 300 -scale 500x5000> -alpha remove -strip -auto-orient /tmp/paperless/paperless-iabw4ryn/archive.pdf[0] /tmp/paperless/paperless-iabw4ryn/convert.webp\",\"[2023-09-08 12:05:27,938] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:05:28,195] [DEBUG] [paperless.consumer] Saving record to database\",\"[2023-09-08 12:05:28,258] [DEBUG] [paperless.consumer] Creation date from st_mtime: 2023-09-08 12:04:53+00:00\",\"[2023-09-08 12:05:29,748] [DEBUG] [paperless.consumer] Deleting file /tmp/paperless/paperless-ngxm31l7wip/test.pdf\",\"[2023-09-08 12:05:29,832] [DEBUG] [paperless.parsing.tesseract] Deleting directory /tmp/paperless/paperless-iabw4ryn\",\"[2023-09-08 12:05:29,836] [INFO] [paperless.consumer] Document 2023-09-08 test consumption finished\",\"[2023-09-08 12:05:37,518] [INFO] [paperless.tasks] No automatic matching items, not training\",\"[2023-09-08 12:06:21,057] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:06:21,069] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:06:22,956] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:06:24,146] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:10:27,642] [DEBUG] [paperless.filehandling] Document has storage_path 1 (testdirp) set\",\"[2023-09-08 12:10:27,852] [DEBUG] [paperless.filehandling] Document has storage_path 1 (testdirp) set\",\"[2023-09-08 12:10:27,874] [DEBUG] [paperless.filehandling] Document has storage_path 1 (testdirp) set\",\"[2023-09-08 12:10:36,578] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:10:43,556] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:10:52,429] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:10:57,501] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:11:10,623] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:14:46,620] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:14:47,230] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:14:55,240] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:15:01,444] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:15:02,965] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:15:15,266] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:15:24,462] [DEBUG] [paperless.filehandling] Document has storage_path 1 (testdirp) set\",\"[2023-09-08 12:15:24,661] [DEBUG] [paperless.filehandling] Document has storage_path 1 (testdirp) set\"]"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "UserInfo" : {
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string",
            "nullable" : true
          },
          "password" : {
            "type" : "string",
            "nullable" : true
          }
        },
        "additionalProperties" : false
      },
...
      "inline_response_200_12" : {
        "required" : [ "all", "count", "next", "previous", "results" ],
        "type" : "object",
        "properties" : {
          "count" : {
            "type" : "integer"
          },
          "next" : {
            "type" : "integer"
          },
          "previous" : {
            "type" : "integer"
          },
          "all" : {
            "type" : "array",
            "items" : { }
          },
          "results" : {
            "type" : "array",
            "items" : { }
          }
        },
        "example" : {
          "next" : 6,
          "all" : [ "", "" ],
          "previous" : 1,
          "count" : 0,
          "results" : [ "", "" ]
        }
      },
      "api_groups_body" : {
        "required" : [ "id", "name", "permissions" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "permissions" : {
            "type" : "array",
            "items" : {
              "required" : [ "name", "permissions" ],
              "type" : "string",
              "properties" : {
                "name" : {
                  "type" : "string"
                },
                "permissions" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      },
      "apigroupsid_permissions_form" : {
        "required" : [ "set_permissions" ],
        "type" : "object",
        "properties" : {
          "set_permissions" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "groups_id_body" : {
        "required" : [ "id", "name", "permissions", "permissions_form", "set_permissions" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "permissions" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "set_permissions" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "permissions_form" : {
            "$ref" : "#/components/schemas/apigroupsid_permissions_form"
          }
        }
      },
      "inline_response_200_13" : {
        "required" : [ "id", "name", "permissions" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "permissions" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "example" : {
          "permissions" : [ "permissions", "permissions" ],
          "name" : "name",
          "id" : 0
        }
      },
...
    }
  }
}